Move frob inner product to math
